+++ src/app/auth/login/page.tsx
@@ -5,7 +5,7 @@ import { AnimatePresence, motion } from "framer-motion";
 import Image from "next/image";
 
 import Onboarding from "@/components/Onboarding";
-import KakaoLoginButton from "@/components/KakaoLoginButton";
+import KakaoLoginButton from "@/components/common/KakaoLoginButton";
 
 import MainGraphic from "/public/img/login_graphic.svg";
